renamed ingester -> worker
