+++ .github/workflows/integration.yml
@@ -52,7 +52,7 @@ jobs:
         mkdir -p data/elastic/
         chmod -R 777 data/
         cp .env.integration .env
-        docker compose up -d --wait core-api ingester
+        docker compose up -d --wait core-api worker
 
     - name: Wait 60s for services to be ready
       run: |
