+++ .github/workflows/python-app.yml
@@ -8,7 +8,7 @@ on:
     paths:
       - redbox/**
       - core_api/**
-      - ingester/**
+      - worker/**
       - Makefile
       - poetry.lock
       - .github/**
@@ -50,10 +50,10 @@ jobs:
         cp .env.test .env
         docker compose up -d --wait elasticsearch
 
-        poetry install --no-root --no-ansi --with dev --without ai,api,ingester
+        poetry install --no-root --no-ansi --with dev --without ai,api,worker
         poetry run download-model --model_name paraphrase-albert-small-v2
 
-    - name: Add dependencies for Ingester
+    - name: Add dependencies for worker
       run: |
         sudo apt update && sudo apt-get install -y \
         poppler-utils \
@@ -63,9 +63,9 @@ jobs:
       run: |
         make test-redbox
 
-    - name: Test ingester with pytest
+    - name: Test worker with pytest
       run: |
-        make test-ingester
+        make test-worker
 
     - name: Test core with pytest
       run: |
