+++ Makefile
@@ -7,7 +7,7 @@ reqs:
 	poetry install
 
 run:
-	docker compose up -d elasticsearch kibana ingester minio redis core-api db django-app
+	docker compose up -d elasticsearch kibana worker minio redis core-api db django-app
 
 stop:
 	docker compose down
@@ -22,16 +22,16 @@ rebuild:
 	docker compose build --no-cache
 
 test-core-api:
-	poetry install --no-root --no-ansi --with api,dev,ai --without ingester
+	poetry install --no-root --no-ansi --with api,dev,ai --without worker
 	poetry run pytest core_api/tests --cov=core_api/src -v --cov-report=term-missing --cov-fail-under=45
 
 test-redbox:
-	poetry install --no-root --no-ansi --with api,dev --without ai,ingester
+	poetry install --no-root --no-ansi --with api,dev --without ai,worker
 	poetry run pytest redbox/tests --cov=redbox -v --cov-report=term-missing --cov-fail-under=45
 
-test-ingester:
-	poetry install --no-root --no-ansi --with ingester,dev --without ai,api
-	poetry run pytest ingester/tests --cov=ingester -v --cov-report=term-missing --cov-fail-under=40
+test-worker:
+	poetry install --no-root --no-ansi --with worker,dev --without ai,api
+	poetry run pytest worker/tests --cov=worker -v --cov-report=term-missing --cov-fail-under=40
 
 test-django:
 	docker compose up -d --wait db minio
@@ -40,9 +40,9 @@ test-django:
 test-integration:
 	docker compose down
 	cp .env.integration .env
-	docker compose build core-api ingester minio
-	docker compose up -d core-api ingester minio
-	poetry install --no-root --no-ansi --with dev --without ai,api,ingester
+	docker compose build core-api worker minio
+	docker compose up -d core-api worker minio
+	poetry install --no-root --no-ansi --with dev --without ai,api,worker
 	sleep 10
 	poetry run pytest tests
 
@@ -51,8 +51,6 @@ lint:
 
 format:
 	poetry run ruff format .
-	# additionally we format, but not lint, the notebooks
-	# poetry run ruff format **/*.ipynb
 
 safe:
 	poetry run bandit -ll -r ./redbox
@@ -61,8 +59,7 @@ safe:
 	poetry run mypy ./django_app --ignore-missing-imports
 
 checktypes:
-	poetry run mypy redbox ingester --ignore-missing-imports
-	# poetry run mypy legacy_app --follow-imports skip --ignore-missing-imports
+	poetry run mypy redbox worker --ignore-missing-imports
 
 check-migrations:
 	docker compose build django-app
