+++ docs/DEVELOPER_SETUP.md
@@ -38,7 +38,7 @@ Typically this involves setting the following variables:
 
 ## Other dependencies (for Document Ingestion and OCR)
 
-You will need to install `poppler` and `tesseract` to run the `ingester`
+You will need to install `poppler` and `tesseract` to run the `worker`
 - `brew install poppler`
 - `brew install tesseract`
 
@@ -52,7 +52,7 @@ The project currently consists of multiple docker images needed to run the proje
 
 - `elasticsearch`
 - `kibana`
-- `ingester`
+- `worker`
 - `minio`
 - `redis`
 - `core-api`
@@ -107,10 +107,10 @@ For the core API:
 make test-core-api
 ```
 
-For the ingester:
+For the worker:
 
 ``` bash
-make test-ingester
+make test-worker
 ```
 
 For integration tests:
