+++ docs/architecture/workers_and_queues.md
@@ -17,14 +17,11 @@ When running locally in docker compose, there is only one replica of each Worker
 
 We are using [FastStream](https://faststream.airt.ai/latest/faststream/) to handle our streaming between Microservices. It handles the connection to Redis (and other Queues) with a high-performance Python client based on FastAPI and Pydantic. It has support for multiple brokers, automatic documentation, and tests. 
 
-## Ingester
-
-The Ingester Worker is responsible for ingesting files into the system. The Ingester Worker reads a [`File`](../code_reference/models/file.md) reference from its queue and then reads the file from the Object Store. The Ingester Worker then processes the file and stores the file in the Database. The Ingester Worker also sends created [`Chunk`](../code_reference/models/chunk.md) references to the Embedder Worker via the `embedding-queue`.
-
-::: ingester.src.worker.ingest
-
 ## Worker
 
-The Worker is responsible for embedding chunks of text. The Embedder Worker reads a [`Chunk`](../code_reference/models/chunk.md) reference from its queue and then reads the text from the Database. The Embedder Worker then embeds the text with it instance of the embedding model. The Embedder Worker then stores the embedding in the Database.
+The Worker is responsible for:
+* ingesting files into the system. The Ingester Worker reads a [`File`](../code_reference/models/file.md) reference from its queue and then reads the file from the Object Store. The Worker then processes the file and stores the file in the Database. The worker also sends created [`Chunk`](../code_reference/models/chunk.md) references to the Worker via the `embedding-queue`.
+* embedding chunks of text. The Embedder Worker reads a [`Chunk`](../code_reference/models/chunk.md) reference from its queue and then reads the text from the Database. The Worker then embeds the text with it instance of the embedding model. The worker then stores the embedding in the Database.
 
-::: embedder.src.worker.embed
+::: worker.src.app.embed
+::: worker.src.app.ingest
