+++ infrastructure/aws/ecs.tf
@@ -64,15 +64,15 @@ module "core_api" {
 }
 
 
-module "ingester" {
+module "worker" {
   create_networking  = false
   memory             = 4096
   cpu                = 2048
   source             = "../../../i-ai-core-infrastructure//modules/ecs"
-  project_name       = "redbox-ingester"
+  project_name       = "redbox-worker"
   image_tag          = var.image_tag
   prefix             = local.prefix
-  ecr_repository_uri = "${var.ecr_repository_uri}/redbox-ingester"
+  ecr_repository_uri = "${var.ecr_repository_uri}/redbox-worker"
   ecs_cluster_id     = module.cluster.ecs_cluster_id
   health_check = {
     healthy_threshold   = 3
